Conflicts:
	build/Makefile_xml
	include/igl/viewer/Viewer.cpp

Former-commit-id: 11ebbb961006aa8f4cef2657e872a4ca1ec17536

+++ include/igl/doublearea.cpp
@@ -132,7 +132,7 @@ IGL_INLINE void igl::doublearea(
   #  define IGL_OMP_MIN_VALUE 1000
   #endif
   #pragma omp parallel for if (m>IGL_OMP_MIN_VALUE)
-  for(size_t i = 0;i<m;i++)
+  for(long int i = 0;i<m;i++)
   {
     //// Heron's formula for area
     //const typename Derivedl::Scalar arg =
